requirements.txt
phidata
python-dotenv
yfinance
duckduckgo-search
fastapi
uvicorn
groq
openai
sqlalchemy
pgvector
psycopg[binary]
pypdf
streamlit
phidata
google-generativeai
